+++ jdaviz/app.py
@@ -304,6 +304,8 @@ class ApplicationState(State):
     # https://github.com/spacetelescope/jdaviz/pull/3835
     # https://github.com/spacetelescope/jdaviz/pull/3854
     # https://github.com/spacetelescope/jdaviz/pull/3856
+    # https://github.com/spacetelescope/jdaviz/pull/3863
+    # https://github.com/spacetelescope/jdaviz/pull/3867 - table viewer
     catalogs_in_dc = CallbackProperty(
         False, docstring="Whether to enable developer mode for adding catalogs to data collection.")
     loader_items = ListCallbackProperty(